Review comments (wording, record tree/quota-user in metrics)
